requirements.txt
# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
websockets>=12.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# ePub Processing
ebooklib>=0.18
beautifulsoup4>=4.12.0
lxml>=4.9.0

# LLM Integration
tiktoken>=0.5.0
openai>=1.0.0
httpx>=0.26.0

# Security
cryptography>=41.0.0
python-jose[cryptography]>=3.3.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Text Processing
# difflib is built-in to Python
Markdown==3.5.2

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
